refactor: decouple prql_parser from compiler Error
